refactor(api): moved verifyUser.js to middleware folder
