+++ api/middleware/verifyUser.js
@@ -1,5 +1,5 @@
 import jwt from "jsonwebtoken";
-import { errorHandler } from "./error.js";
+import { errorHandler } from "../utils/error.js";
 export const verifyToken = (req, res, next) => {
    const token = req.cookies.access_token;
