+++ api/routes/admin.route.js
@@ -1,6 +1,6 @@
 import express from "express";
 import { getUsers, deleteUser, getListings, deleteListing } from "../controllers/admin.controller.js";
-import { verifyToken } from "../utils/verifyUser.js";
+import { verifyToken } from "../middleware/verifyUser.js";
 import verifyRoles from "../middleware/verifyRoles.js";
 
 const router = express.Router();
